backtracking mas legible

+++ ejercicioEntregable_2/Trabajo_Practico_Entregable/src/ejercicioEntregable/modelo/Alumno.java
@@ -61,7 +61,7 @@ public class Alumno implements Comparable<Alumno> {
 	}
 	
 	public boolean estaAprobado(float nota) {	
-		return puntaje >= nota;
+		return puntaje >= (int) nota;
 	}
 	
 	@Override
